Move DatasourceScheduleType to datasources struct Remove some unused imports in datasource controller Fix updateDatasource call creating a nested object with the connection body under "datasource" Hide some tabs and fix the schedule editing tab Remove BASICSCHEDULE fron frontend because airbyte new API doesn't support it yet (its just cron anyway) close #264
