--- webapp/src/lib/struct/schedule.ts
@@ -1,5 +1,0 @@
-export enum DatasourceScheduleType {
-	MANUAL = 'manual',
-	BASICSCHEDULE = 'basic',
-	CRON = 'cron',
-};
